requirements.txt
telebot
beautifulsoup4
requests
selenium
pytesseract
Pillow
webdriver-manager